+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- x-axis limits for accuracy, mean read quality, and read alignment coverage histograms to be more dynamic.
+
 ## [v0.5.2]
 ### Changed
 - Removed no longer used `--concat_fastq` parameter.