+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Fixed
+- Read length histogram only displaying a small number of bins when there are a few outlier reads a lot longer than the other reads.
+
 ### Changed
 - x-axis limits for accuracy, mean read quality, and read alignment coverage histograms to be more dynamic.
 ### Fixed