+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [v0.5.3]
 ### Fixed
 - Read length histogram only displaying a small number of bins when there are a few outlier reads a lot longer than the other reads.
+- configure-jbrowse breaking on unescaped spaces
 
 ### Changed
 - x-axis limits for accuracy, mean read quality, and read alignment coverage histograms to be more dynamic.